# api/swagger.yaml
...
swagger: "2.0"
info:
  description: "Ynov API for YnovImmo"
  version: "1.0.0"
  title: "Ynov API pour l'application YnovImmo"
host: "localhost:8080"
basePath: "/api/v1"
schemes:
- "http"
paths:
...
  /real-estate:
    get:
      tags:
      - "real-estate"
      summary: "get all real-estate by pagination"
      description: ""
      operationId: "real_estateGET"
      consumes:
      - "application/json"
      produces:
      - "application/json"
      parameters:
      - name: "where"
        in: "query"
        description: "column:value will use sql LIKE for search eg:id:67 will where\
          \ id >67 eg2: name:eric => where name LIKE '%eric%'"
        required: false
        type: "array"
        items:
          type: "string"
        collectionFormat: "multi"
      - name: "fields"
        in: "query"
        description: "{$tableColumn},{$tableColumn}... "
        required: false
        type: "string"
      - name: "order"
        in: "query"
        description: "eg: id desc, name desc"
        required: false
        type: "string"
      - name: "offset"
        in: "query"
        description: "sql offset eg: 10"
        required: false
        type: "integer"
      - name: "limit"
        in: "query"
        description: "limit returning object count"
        required: false
        type: "integer"
        default: 2
      responses:
        "200":
          description: "successful operation"
          schema:
            type: "array"
            items:
              $ref: "#/definitions/RealEstatePagination"
      x-swagger-router-controller: "RealEstate"
    post:
      tags:
      - "real-estate"
      summary: "create real-estate"
      description: "create real-estate"
      operationId: "real_estatePOST"
      consumes:
      - "application/json"
      produces:
      - "application/json"
      parameters:
      - in: "body"
        name: "body"
        description: "create real-estate"
        required: true
        schema:
          $ref: "#/definitions/RealEstate"
      responses:
        "200":
          description: "successful operation"
          schema:
            $ref: "#/definitions/ApiResponse"
      x-swagger-router-controller: "RealEstate"
    patch:
      tags:
      - "real-estate"
      summary: "update real-estate"
      description: "update real-estate"
      operationId: "real_estatePATCH"
      consumes:
      - "application/json"
      produces:
      - "application/json"
      parameters:
      - in: "body"
        name: "body"
        description: "create real-estate"
        required: true
        schema:
          $ref: "#/definitions/RealEstate"
      responses:
        "200":
          description: "successful operation"
          schema:
            $ref: "#/definitions/ApiResponse"
      x-swagger-router-controller: "RealEstate"
  /real-estate/{Id}:
    get:
      tags:
      - "real-estate"
      summary: "get a real-estate by ID"
      description: "get a real-estate by ID"
      operationId: "real_estateIdGET"
      consumes:
      - "application/json"
      produces:
      - "application/json"
      parameters:
      - name: "Id"
        in: "path"
        description: "ID of real-estate to return"
        required: true
        type: "integer"
        format: "int64"
      responses:
        "200":
          description: "successful operation"
          schema:
            $ref: "#/definitions/RealEstate"
      x-swagger-router-controller: "RealEstate"
    delete:
      tags:
      - "real-estate"
      summary: "Destroy a real-estate by ID"
      description: "delete a real-estate by ID"
      operationId: "real_estateIdDELETE"
      consumes:
      - "application/json"
      produces:
      - "application/json"
      parameters:
      - name: "Id"
        in: "path"
        description: "ID of real-estate to delete"
        required: true
        type: "integer"
        format: "int64"
      responses:
        "200":
          description: "successful operation"
          schema:
            $ref: "#/definitions/ApiResponse"
      x-swagger-router-controller: "RealEstate"
  /real-estate-availability:
    get:
      tags:
      - "real-estate-availability"
      summary: "get all real-estate-availability by pagination"
      description: ""
      operationId: "real_estate_availabilityGET"
      consumes:
      - "application/json"
      produces:
      - "application/json"
      parameters:
      - name: "where"
        in: "query"
        description: "column:value will use sql LIKE for search eg:id:67 will where\
          \ id >67 eg2: name:eric => where name LIKE '%eric%'"
        required: false
        type: "array"
        items:
          type: "string"
        collectionFormat: "multi"
      - name: "fields"
        in: "query"
        description: "{$tableColumn},{$tableColumn}... "
        required: false
        type: "string"
      - name: "order"
        in: "query"
        description: "eg: id desc, name desc"
        required: false
        type: "string"
      - name: "offset"
        in: "query"
        description: "sql offset eg: 10"
        required: false
        type: "integer"
      - name: "limit"
        in: "query"
        description: "limit returning object count"
        required: false
        type: "integer"
        default: 2
      responses:
        "200":
          description: "successful operation"
          schema:
            type: "array"
            items:
              $ref: "#/definitions/RealEstateAvailabilityPagination"
      x-swagger-router-controller: "RealEstateAvailability"
    post:
      tags:
      - "real-estate-availability"
      summary: "create real-estate-availability"
      description: "create real-estate-availability"
      operationId: "real_estate_availabilityPOST"
      consumes:
      - "application/json"
      produces:
      - "application/json"
      parameters:
      - in: "body"
        name: "body"
        description: "create real-estate-availability"
        required: true
        schema:
          $ref: "#/definitions/RealEstateAvailability"
      responses:
        "200":
          description: "successful operation"
          schema:
            $ref: "#/definitions/ApiResponse"
      x-swagger-router-controller: "RealEstateAvailability"
    patch:
      tags:
      - "real-estate-availability"
      summary: "update real-estate-availability"
      description: "update real-estate-availability"
      operationId: "real_estate_availabilityPATCH"
      consumes:
      - "application/json"
      produces:
      - "application/json"
      parameters:
      - in: "body"
        name: "body"
        description: "create real-estate-availability"
        required: true
        schema:
          $ref: "#/definitions/RealEstateAvailability"
      responses:
        "200":
          description: "successful operation"
          schema:
            $ref: "#/definitions/ApiResponse"
      x-swagger-router-controller: "RealEstateAvailability"
  /real-estate-availability/{Id}:
    get:
      tags:
      - "real-estate-availability"
      summary: "get a real-estate-availability by ID"
      description: "get a real-estate-availability by ID"
      operationId: "real_estate_availabilityIdGET"
      consumes:
      - "application/json"
      produces:
      - "application/json"
      parameters:
      - name: "Id"
        in: "path"
        description: "ID of real-estate-availability to return"
        required: true
        type: "integer"
        format: "int64"
      responses:
        "200":
          description: "successful operation"
          schema:
            $ref: "#/definitions/RealEstateAvailability"
      x-swagger-router-controller: "RealEstateAvailability"
    delete:
      tags:
      - "real-estate-availability"
      summary: "Destroy a real-estate-availability by ID"
      description: "delete a real-estate-availability by ID"
      operationId: "real_estate_availabilityIdDELETE"
      consumes:
      - "application/json"
      produces:
      - "application/json"
      parameters:
      - name: "Id"
        in: "path"
        description: "ID of real-estate-availability to delete"
        required: true
        type: "integer"
        format: "int64"
      responses:
        "200":
          description: "successful operation"
          schema:
            $ref: "#/definitions/ApiResponse"
      x-swagger-router-controller: "RealEstateAvailability"
  /real-estate-feedback:
    get:
      tags:
      - "real-estate-feedback"
      summary: "get all real-estate-feedback by pagination"
      description: ""
      operationId: "real_estate_feedbackGET"
      consumes:
      - "application/json"
      produces:
      - "application/json"
      parameters:
      - name: "where"
        in: "query"
        description: "column:value will use sql LIKE for search eg:id:67 will where\
          \ id >67 eg2: name:eric => where name LIKE '%eric%'"
        required: false
        type: "array"
        items:
          type: "string"
        collectionFormat: "multi"
      - name: "fields"
        in: "query"
        description: "{$tableColumn},{$tableColumn}... "
        required: false
        type: "string"
      - name: "order"
        in: "query"
        description: "eg: id desc, name desc"
        required: false
        type: "string"
      - name: "offset"
        in: "query"
        description: "sql offset eg: 10"
        required: false
        type: "integer"
      - name: "limit"
        in: "query"
        description: "limit returning object count"
        required: false
        type: "integer"
        default: 2
      responses:
        "200":
          description: "successful operation"
          schema:
            type: "array"
            items:
              $ref: "#/definitions/RealEstateFeedbackPagination"
      x-swagger-router-controller: "RealEstateFeedback"
    post:
      tags:
      - "real-estate-feedback"
      summary: "create real-estate-feedback"
      description: "create real-estate-feedback"
      operationId: "real_estate_feedbackPOST"
      consumes:
      - "application/json"
      produces:
      - "application/json"
      parameters:
      - in: "body"
        name: "body"
        description: "create real-estate-feedback"
        required: true
        schema:
          $ref: "#/definitions/RealEstateFeedback"
      responses:
        "200":
          description: "successful operation"
          schema:
            $ref: "#/definitions/ApiResponse"
      x-swagger-router-controller: "RealEstateFeedback"
    patch:
      tags:
      - "real-estate-feedback"
      summary: "update real-estate-feedback"
      description: "update real-estate-feedback"
      operationId: "real_estate_feedbackPATCH"
      consumes:
      - "application/json"
      produces:
      - "application/json"
      parameters:
      - in: "body"
        name: "body"
        description: "create real-estate-feedback"
        required: true
        schema:
          $ref: "#/definitions/RealEstateFeedback"
      responses:
        "200":
          description: "successful operation"
          schema:
            $ref: "#/definitions/ApiResponse"
      x-swagger-router-controller: "RealEstateFeedback"
  /real-estate-feedback/{Id}:
    get:
      tags:
      - "real-estate-feedback"
      summary: "get a real-estate-feedback by ID"
      description: "get a real-estate-feedback by ID"
      operationId: "real_estate_feedbackIdGET"
      consumes:
      - "application/json"
      produces:
      - "application/json"
      parameters:
      - name: "Id"
        in: "path"
        description: "ID of real-estate-feedback to return"
        required: true
        type: "integer"
        format: "int64"
      responses:
        "200":
          description: "successful operation"
          schema:
            $ref: "#/definitions/RealEstateFeedback"
      x-swagger-router-controller: "RealEstateFeedback"
    delete:
      tags:
      - "real-estate-feedback"
      summary: "Destroy a real-estate-feedback by ID"
      description: "delete a real-estate-feedback by ID"
      operationId: "real_estate_feedbackIdDELETE"
      consumes:
      - "application/json"
      produces:
      - "application/json"
      parameters:
      - name: "Id"
        in: "path"
        description: "ID of real-estate-feedback to delete"
        required: true
        type: "integer"
        format: "int64"
      responses:
        "200":
          description: "successful operation"
          schema:
            $ref: "#/definitions/ApiResponse"
      x-swagger-router-controller: "RealEstateFeedback"
...
definitions:
...
  RealEstate:
    type: "object"
    properties:
      id:
        type: "integer"
        format: "int64"
        description: "Id du bien"
      id_user:
        type: "integer"
        format: "int64"
        description: "Id du proprietaire"
      accroche:
        type: "string"
        format: "string"
        description: "Phrase d'accroche (max 100 caractères)"
      type:
        type: "string"
        format: "string"
        description: "Type de bien ('apartment','house','vacant_lot','parking','loft','castle','building')"
      nb_rooms:
        type: "integer"
        format: "int64"
        description: "Nombre de pièce"
      nb_bedroom:
        type: "integer"
        format: "int64"
        description: "Nombre de chambre"
      description:
        type: "string"
        format: "string"
        description: "Description"
      size:
        type: "integer"
        format: "int64"
        description: "Superficie (en entier)"
      price:
        type: "integer"
        format: "int64"
        description: "Prix du bien"
      address:
        type: "string"
        format: "string"
        description: "Adresse"
      zip_code:
        type: "string"
        format: "string"
        description: "Code postal"
      city:
        type: "string"
        format: "string"
        description: "Ville"
      latitude:
        type: "string"
        format: "string"
        description: "Latitude"
      longitude:
        type: "string"
        format: "string"
        description: "Longitude"
      energy_class:
        type: "string"
        format: "string"
        description: "Classe d'énergie"
      ges_class:
        type: "string"
        format: "string"
        description: "Classe GES"
      has_garden:
        type: "integer"
        format: "int64"
        description: "A un jardin"
      has_exposed_stone:
        type: "integer"
        format: "int64"
        description: "A des pierres apparentes"
      has_ciment_tiles:
        type: "integer"
        format: "int64"
        description: "A des carreaux de ciments"
      has_parquet_floor:
        type: "integer"
        format: "int64"
        description: "A du parquet au sol"
    example:
      ges_class: "ges_class"
      address: "address"
      city: "city"
      accroche: "accroche"
      latitude: "latitude"
      description: "description"
      id_user: 2
      type: "type"
      zip_code: "zip_code"
      has_exposed_stone: 7
      has_garden: 4
      nb_bedroom: 9
      size: 3
      price: 2
      nb_rooms: 7
      energy_class: "energy_class"
      id: 5
      has_parquet_floor: 1
      longitude: "longitude"
      has_ciment_tiles: 1
  RealEstatePagination:
    type: "object"
    properties:
      code:
        type: "integer"
        format: "int32"
        description: "json repose code"
      total:
        type: "integer"
        format: "int32"
        description: "total numbers"
      offset:
        type: "integer"
        format: "int32"
        description: "offset"
      limit:
        type: "integer"
        format: "int32"
        description: "limit"
      list:
        type: "array"
        items:
          $ref: "#/definitions/RealEstate"
    example:
      total: 6
      code: 0
      offset: 1
      limit: 5
      list:
      - ges_class: "ges_class"
        address: "address"
        city: "city"
        accroche: "accroche"
        latitude: "latitude"
        description: "description"
        id_user: 2
        type: "type"
        zip_code: "zip_code"
        has_exposed_stone: 7
        has_garden: 4
        nb_bedroom: 9
        size: 3
        price: 2
        nb_rooms: 7
        energy_class: "energy_class"
        id: 5
        has_parquet_floor: 1
        longitude: "longitude"
        has_ciment_tiles: 1
      - ges_class: "ges_class"
        address: "address"
        city: "city"
        accroche: "accroche"
        latitude: "latitude"
        description: "description"
        id_user: 2
        type: "type"
        zip_code: "zip_code"
        has_exposed_stone: 7
        has_garden: 4
        nb_bedroom: 9
        size: 3
        price: 2
        nb_rooms: 7
        energy_class: "energy_class"
        id: 5
        has_parquet_floor: 1
        longitude: "longitude"
        has_ciment_tiles: 1
  RealEstateAvailability:
    type: "object"
    properties:
      id:
        type: "integer"
        format: "int64"
        description: "Id de la disponibilité"
      price:
        type: "number"
        format: "float"
        description: "Prix de la visite"
      start_date:
        type: "string"
        format: "date-time"
        description: "Date de début"
      end_date:
        type: "string"
        format: "date-time"
        description: "Date de fin"
      start_time:
        type: "string"
        format: "date-time"
        description: "De "
      end_time:
        type: "string"
        format: "date-time"
        description: "à"
    example:
      end_date: "2000-01-23T04:56:07.000+00:00"
      start_time: "2000-01-23T04:56:07.000+00:00"
      price: 2.302136
      end_time: "2000-01-23T04:56:07.000+00:00"
      id: 5
      start_date: "2000-01-23T04:56:07.000+00:00"
  RealEstateAvailabilityPagination:
    type: "object"
    properties:
      code:
        type: "integer"
        format: "int32"
        description: "json repose code"
      total:
        type: "integer"
        format: "int32"
        description: "total numbers"
      offset:
        type: "integer"
        format: "int32"
        description: "offset"
      limit:
        type: "integer"
        format: "int32"
        description: "limit"
      list:
        type: "array"
        items:
          $ref: "#/definitions/RealEstateAvailability"
    example:
      total: 6
      code: 0
      offset: 1
      limit: 5
      list:
      - end_date: "2000-01-23T04:56:07.000+00:00"
        start_time: "2000-01-23T04:56:07.000+00:00"
        price: 2.302136
        end_time: "2000-01-23T04:56:07.000+00:00"
        id: 5
        start_date: "2000-01-23T04:56:07.000+00:00"
      - end_date: "2000-01-23T04:56:07.000+00:00"
        start_time: "2000-01-23T04:56:07.000+00:00"
        price: 2.302136
        end_time: "2000-01-23T04:56:07.000+00:00"
        id: 5
        start_date: "2000-01-23T04:56:07.000+00:00"
  RealEstateFeedback:
    type: "object"
    properties:
      id:
        type: "integer"
        format: "int64"
        description: "Id du feedback"
      id_user:
        type: "integer"
        format: "int64"
        description: "Id de l'utilisateur"
      id_real_estate:
        type: "integer"
        format: "int64"
        description: "Id du bien"
      feedback:
        type: "string"
        format: "string"
        description: "Feedback"
      date:
        type: "string"
        format: "date-time"
        description: "Date (YYYY-mm-dd hh:mm:ss)"
    example:
      feedback: "feedback"
      date: "2000-01-23T04:56:07.000+00:00"
      id: 5
      id_user: 2
      id_real_estate: 7
  RealEstateFeedbackPagination:
    type: "object"
    properties:
      code:
        type: "integer"
        format: "int32"
        description: "json repose code"
      total:
        type: "integer"
        format: "int32"
        description: "total numbers"
      offset:
        type: "integer"
        format: "int32"
        description: "offset"
      limit:
        type: "integer"
        format: "int32"
        description: "limit"
      list:
        type: "array"
        items:
          $ref: "#/definitions/RealEstateFeedback"
    example:
      total: 6
      code: 0
      offset: 1
      limit: 5
      list:
      - feedback: "feedback"
        date: "2000-01-23T04:56:07.000+00:00"
        id: 5
        id_user: 2
        id_real_estate: 7
      - feedback: "feedback"
        date: "2000-01-23T04:56:07.000+00:00"
        id: 5
        id_user: 2
        id_real_estate: 7
...
  ApiResponse:
    type: "object"
    properties:
      code:
        type: "integer"
        format: "int32"
      msg:
        type: "string"
    example:
      msg: "msg"
      code: 0
...